requirements.txt
anyio==3.7.0
certifi==2023.5.7
cryptography==41.0.0                  
dnspython==2.3.0
databases
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.11
orjson==3.8.14
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
protobuf==3.20.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pyflakes==3.0.1
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0